Updated to support emulation timing with picopass + cleanup
