+++ application.fam
@@ -8,7 +8,7 @@ App(
     stack_size=4 * 1024,  # Increased for picopass emulation
     fap_category="Tools",
     # Optional values
-    fap_version="0.9.1",
+    fap_version="0.10.0",
     fap_icon="images/quac.png",  # 10x10 1-bit PNG
     fap_description="Quick Action remote control app",
     fap_author="Roberto De Feo",
