+++ quac.h
@@ -61,6 +61,7 @@ typedef struct App {
         uint32_t rfid_duration; // Defaults to 2500 ms
         uint32_t nfc_duration; // Defaults to 1000 ms
         uint32_t ibutton_duration; // Defaults to 1000 ms
+        uint32_t picopass_duration; // Defaults to 1000 ms (used on tap & as playlist default)
         bool ir_use_ext_module; // Defaults to False
         bool show_hidden; // Defaults to False
     } settings;
