--- scenes/scene_picopass_emulate.h
@@ -1,5 +1,0 @@
-#pragma once
-
-void scene_picopass_emulate_on_enter(void* context);
-bool scene_picopass_emulate_on_event(void* context, SceneManagerEvent event);
-void scene_picopass_emulate_on_exit(void* context);
