+++ scenes/scenes.c
@@ -9,7 +9,6 @@
 #include "scene_action_create_group.h"
 #include "scene_action_ir_list.h"
 #include "scene_about.h"
-#include "scene_picopass_emulate.h"
 
 // define handler callbacks - order must match appScenes enum!
 void (*const app_on_enter_handlers[])(void* context) = {
@@ -20,7 +19,6 @@ void (*const app_on_enter_handlers[])(void* context) = {
     scene_action_create_group_on_enter,
     scene_action_ir_list_on_enter,
     scene_about_on_enter,
-    scene_picopass_emulate_on_enter,
 };
 bool (*const app_on_event_handlers[])(void* context, SceneManagerEvent event) = {
     scene_items_on_event,
@@ -30,7 +28,6 @@ bool (*const app_on_event_handlers[])(void* context, SceneManagerEvent event) = 
     scene_action_create_group_on_event,
     scene_action_ir_list_on_event,
     scene_about_on_event,
-    scene_picopass_emulate_on_event,
 };
 void (*const app_on_exit_handlers[])(void* context) = {
     scene_items_on_exit,
@@ -40,7 +37,6 @@ void (*const app_on_exit_handlers[])(void* context) = {
     scene_action_create_group_on_exit,
     scene_action_ir_list_on_exit,
     scene_about_on_exit,
-    scene_picopass_emulate_on_exit,
 };
 
 const SceneManagerHandlers app_scene_handlers = {
