+++ scenes/scenes.h
@@ -9,7 +9,6 @@ typedef enum {
     QScene_ActionCreateGroup,
     QScene_ActionIRList,
     QScene_About,
-    QScene_PicopassEmulate,
     QScene_count
 } appScenes;
